🐛 QuizbookCard 오류 수정
