+++ src/components/QuizbookCard/index.ts
@@ -1,0 +1,16 @@
+import QuizbookCardRoot from './QuizbookCardRoot'
+import Description from './Description'
+import Author from './Author'
+import SolvedRate from './SolvedRate'
+import ReviewRate from './ReviewRate'
+import QuizCount from './QuizCount'
+import LikeButton from './LikeButton'
+
+export default Object.assign(QuizbookCardRoot, {
+    Description,
+    Author,
+    SolvedRate,
+    ReviewRate,
+    QuizCount,
+    LikeButton,
+})
